pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel>=0.41.0"]
build-backend = "setuptools.build_meta"

[project]
name = "python-codemode"
version = "0.1.0"
description = "Python sandbox execution with LLM code generation - a Python port of Cloudflare's Codemode"
readme = "docs/README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "python-codemode contributors"},
]
keywords = ["sandbox", "llm", "code-generation", "codemode", "ai", "mcp", "langchain"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "openai>=1.0.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
langchain = [
    "langchain>=1.0.0",
    "langchain-openai>=0.1.0",
    "langchain-mcp-adapters>=0.1.0",
]
docker = ["docker>=7.0.0"]
all = [
    "python-codemode[langchain]",
    "python-codemode[docker]",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/python-codemode/python-codemode"
Repository = "https://github.com/python-codemode/python-codemode"
Issues = "https://github.com/python-codemode/python-codemode/issues"

[tool.setuptools.packages.find]
include = ["python_codemode*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["python_codemode"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
fail_under = 60
